+++ src/dump.c
@@ -1043,7 +1043,7 @@ sym_operator_name(const char *sym_name, mrb_int len)
   for (start = 0; table_size != 0; table_size/=2) {
     idx = start+table_size/2;
     op_sym = &operator_table[idx];
-    cmp = len-op_sym->sym_name_len;
+    cmp = (int)(len-op_sym->sym_name_len);
     if (cmp == 0) {
       cmp = memcmp(sym_name, op_sym->sym_name, len);
       if (cmp == 0) return op_sym->name;
